requirements.txt
backcall==0.2.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==3.1
decorator==4.4.2
firebase==3.0.1
gcloud==0.18.3
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
jupyter-client==6.1.7
jupyter-core==4.6.3
jwcrypto==0.8
oauth2client==4.1.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.7
protobuf==3.13.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.9.8
Pygments==2.7.1
python-dateutil==2.8.1
python-jwt==3.3.0
pyzmq==19.0.2
requests==2.24.0
requests-toolbelt==0.9.1
rsa==4.6
six==1.15.0
sseclient==0.0.26
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.10
wcwidth==0.2.5
flask
flask_cors
pydicom
pytz
python-docx

